fix low contrast totalaveage.tif generation
In order to save a .tif file a double must be converted using `uint16()`. This process essentially eliminates decimal values, but there was real information in those decimals values in averaged double in Matlab. The double must be multiplied to expand the range of values it exists on to maintain information after applying the conversion.